fix interface notifikasi password salah + delete confirmation

+++ edit_banner.php
@@ -122,7 +122,7 @@ while ($data = mysqli_fetch_array($result)) {
 
                 <div class="tomboledit">
                     <button type="submit" class="simpan-btn" name="bupdate">Simpan</button>
-                    <span class="data-action"> <a href="hapus_banner.php?id_banner=<?php echo $id_banner; ?>">
+                    <span class="data-action"> <a href="hapus_banner.php?id_banner=<?php echo $id_banner; ?>" onclick="return confirm('Apakah anda yakin mau menghapus banner ini?')">
                             <button type="button" class="btnhapus" data-toggle="modal" data-target="#exampleModal">Hapus produk</button>
                         </a></span>
                     <button type="submit" class="kembali-btn" data-dismiss="modal" aria-label="Close">
